+++ src/main/java/com/kustacks/kuring/user/business/UserService.java
@@ -1,5 +1,6 @@
 package com.kustacks.kuring.user.business;
 
+import com.kustacks.kuring.admin.common.dto.FeedbackDto;
 import com.kustacks.kuring.common.exception.NotFoundException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
 import com.kustacks.kuring.message.firebase.FirebaseService;
@@ -11,12 +12,12 @@ import com.kustacks.kuring.user.domain.User;
 import com.kustacks.kuring.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static com.kustacks.kuring.message.firebase.FirebaseService.ALL_DEVICE_SUBSCRIBED_TOPIC;
 
@@ -46,6 +47,17 @@ public class UserService {
     public List<CategoryName> lookUpUserCategories(String token) {
         User findUser = findUserByToken(token);
         return findUser.getSubscribedCategoryList();
+    }
+
+    @Transactional(readOnly = true)
+    public List<FeedbackDto> lookupFeedbacks(int page, int size) {
+        PageRequest pageRequest = PageRequest.of(page, size);
+        return userRepository.findAllFeedbackByPageRequest(pageRequest);
+    }
+
+    public void saveFeedback(String token, String content) {
+        User findUser = findUserByToken(token);
+        findUser.addFeedback(content);
     }
 
     public SubscribeCompareResultDto<CategoryName> editSubscribeCategoryList(
