+++ src/main/java/com/kustacks/kuring/user/facade/UserQueryFacade.java
@@ -1,11 +1,10 @@
 package com.kustacks.kuring.user.facade;
 
 import com.kustacks.kuring.admin.common.dto.FeedbackDto;
-import com.kustacks.kuring.user.business.FeedbackService;
-import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.message.firebase.FirebaseService;
 import com.kustacks.kuring.notice.common.dto.CategoryNameDto;
 import com.kustacks.kuring.notice.common.dto.DepartmentNameDto;
+import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.notice.domain.DepartmentName;
 import com.kustacks.kuring.user.business.UserService;
 import lombok.RequiredArgsConstructor;
@@ -13,14 +12,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class UserQueryFacade {
 
-    private final FeedbackService feedbackService;
     private final UserService userService;
     private final FirebaseService firebaseService;
 
@@ -35,7 +32,7 @@ public class UserQueryFacade {
     }
 
     public List<FeedbackDto> lookupFeedbacks(int page, int size) {
-        return feedbackService.lookupFeedbacks(page, size);
+        return userService.lookupFeedbacks(page, size);
     }
 
     private List<CategoryNameDto> convertCategoryNameDtoList(List<CategoryName> categoryNamesList) {
